updates to support custom image path in operator (#9840)
- update template for pkg/components/image.go to allow setting image path
    - get image path from product registry and operator image for enterprise and tigera operator respectively
    - update pinned component list generated for operator
    - fix code generation check: use the set operator org, repo & branch
    - fix manifest generation
    - update tests

+++ release/pkg/manager/operator/manager_enterprise.go
@@ -15,7 +15,7 @@ import (
 	"github.com/projectcalico/calico/release/internal/registry"
 )
 
-//go:embed template/enterprise-images.go.gotmpl
+//go:embed templates/enterprise-images.go.gotmpl
 var enterpriseComponentImagesFileTemplate string
 
 type EnterpriseOperatorManager struct {
@@ -35,9 +35,6 @@ func (o *EnterpriseOperatorManager) PreBuildValidation() error {
 	if err := o.OperatorManager.PreBuildValidation(o.tmpDir); err != nil {
 		return err
 	}
-	if !strings.HasSuffix(o.productRegistry, fmt.Sprintf("/%s", registry.TigeraNamespace)) {
-		return fmt.Errorf("operator does not support product registry %s, it must end with /%s", o.productRegistry, registry.TigeraNamespace)
-	}
 	return nil
 }
 
@@ -51,6 +48,9 @@ func (o *EnterpriseOperatorManager) Build() error {
 	if err != nil {
 		return err
 	}
+	if component.Image != o.image {
+		return fmt.Errorf("operator image mismatch: expected %s, got %s", o.image, component.Image)
+	}
 	defer func() {
 		if _, err := o.runner.RunInDir(o.dir, "git", []string{"reset", "--hard"}, nil); err != nil {
 			logrus.WithError(err).Error("Failed to reset repository")
@@ -61,23 +61,22 @@ func (o *EnterpriseOperatorManager) Build() error {
 	}
 	env := os.Environ()
 	env = append(env, fmt.Sprintf("EE_VERSIONS=%s", componentsVersionPath))
-	env = append(env, fmt.Sprintf("COMMON_VERSIONS=%s", componentsVersionPath))
 	env = append(env, fmt.Sprintf("ENTERPRISE_CRDS_DIR=%s", o.calicoDir))
 	if _, err := o.make("gen-versions", env); err != nil {
-		return err
+		return fmt.Errorf("failed to generate versions: %w", err)
 	}
 	env = os.Environ()
 	env = append(env, fmt.Sprintf("ARCHES=%s", strings.Join(o.architectures, " ")))
 	env = append(env, fmt.Sprintf("GIT_VERSION=%s", component.Version))
 	env = append(env, fmt.Sprintf("BUILD_IMAGE=%s", component.Image))
 	if _, err := o.make("image-all", env); err != nil {
-		return err
+		return fmt.Errorf("failed to build images: %w", err)
 	}
 	for _, arch := range o.architectures {
 		currentTag := fmt.Sprintf("%s:latest-%s", component.Image, arch)
 		newTag := fmt.Sprintf("%s-%s", component.String(), arch)
 		if err := o.docker.TagImage(currentTag, newTag); err != nil {
-			return err
+			return fmt.Errorf("failed to tag image %q to %q: %w", currentTag, newTag, err)
 		}
 	}
 	env = os.Environ()
@@ -85,11 +84,14 @@ func (o *EnterpriseOperatorManager) Build() error {
 	env = append(env, fmt.Sprintf("BUILD_IMAGE=%s", component.Image))
 	env = append(env, fmt.Sprintf("BUILD_INIT_IMAGE=%s", component.InitImage().Image))
 	if _, err := o.make("image-init", env); err != nil {
-		return err
+		return fmt.Errorf("failed to build init image: %w", err)
 	}
 	currentTag := fmt.Sprintf("%s:latest", component.InitImage().Image)
 	newTag := component.InitImage().String()
-	return o.docker.TagImage(currentTag, newTag)
+	if err := o.docker.TagImage(currentTag, newTag); err != nil {
+		return fmt.Errorf("failed to tag image %q to %q: %w", currentTag, newTag, err)
+	}
+	return nil
 }
 
 // modifyComponentsImagesFile overwrites the pkg/components/images.go file
@@ -100,22 +102,33 @@ func (o *EnterpriseOperatorManager) modifyComponentsImagesFile() error {
 	dest, err := os.OpenFile(destFilePath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0o644)
 	if err != nil {
 		logrus.WithError(err).Errorf("Failed to open file %s", destFilePath)
-		return err
+		return fmt.Errorf("failed to open file %s: %w", destFilePath, err)
 	}
 	defer func() { _ = dest.Close() }()
 	tmpl, err := template.New("pkg/components/images.go").Parse(enterpriseComponentImagesFileTemplate)
 	if err != nil {
 		logrus.WithError(err).Errorf("Failed to parse template to overwrite %s file", destFilePath)
+		return fmt.Errorf("failed to parse template to overwrite %s file: %w", destFilePath, err)
+	}
+
+	imagePath, _, err := o.imageParts()
+	if err != nil {
+		return err
+	}
+	productRegistry, productImagePath, err := o.productRegistryParts()
+	if err != nil {
 		return err
 	}
 
 	if err := tmpl.Execute(dest, map[string]string{
-		"Registry":        o.registry,
-		"ProductRegistry": strings.TrimSuffix(o.productRegistry, fmt.Sprintf("/%s", registry.TigeraNamespace)),
-		"Year":            time.Now().Format("2006"),
+		"ImagePath":        imagePath,
+		"Registry":         o.registry,
+		"ProductImagePath": productImagePath,
+		"ProductRegistry":  productRegistry,
+		"Year":             time.Now().Format("2006"),
 	}); err != nil {
 		logrus.WithError(err).Errorf("Failed to write to file %s", destFilePath)
-		return err
+		return fmt.Errorf("failed to write to file %s: %w", destFilePath, err)
 	}
 	return nil
 }
